+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `%ntbl push datasets [PATH]` command
 - `%ntbl pull datasets [PATH]` command
+- Logging to `/var/log/noteable_magics.log` or locally to `/tmp/noteable_magics.log`
 
 ## [1.1.0] - 2021-06-14
 ### Changed